pyproject.toml
[project]
name = "code-testcase-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "shinkeonkim",email = "dev.shinkeonkim@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "grpcio",
    "grpcio-tools",
    "httpx",
    "pydantic",
    "openai",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "protobuf (>=6.31.1,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
